add additional error checking when people misuse `json_config`
Basically I have seen many users assume this is a filename when in fact it is a key in the sensu settings hash which by default combines all the `*.json` files in `/etc/sensu/conf.d/` and it's subdirs.

Also updated description of `json_config` to make it clearer as to its intent.

Signed-off-by: Ben Abrams <me@benabrams.it>